Move EEP quant method staging to standalone helper
