+++ vllm/model_executor/layers/fused_moe/fused_moe_method_base.py
@@ -2,8 +2,7 @@
 # SPDX-FileCopyrightText: Copyright contributors to the vLLM project
 
 from abc import abstractmethod
-from inspect import signature
-from typing import TYPE_CHECKING, Any
+from typing import TYPE_CHECKING
 
 import torch
 
@@ -107,112 +106,13 @@ class FusedMoEMethodBase(QuantizeMethodBase):
         self,
         routing_tables: tuple[torch.Tensor, torch.Tensor, torch.Tensor] | None = None,
     ) -> FusedMoEPrepareAndFinalizeModular | None:
-        pf = self._make_prepare_finalize(routing_tables=routing_tables)
-        assert pf is None or isinstance(pf, FusedMoEPrepareAndFinalizeModular)
-        return pf
-
-    def _make_prepare_finalize(
-        self,
-        routing_tables: tuple[torch.Tensor, torch.Tensor, torch.Tensor] | None = None,
-        moe_config: FusedMoEConfig | None = None,
-        *,
-        eep_stage: bool = False,
-    ) -> mk.FusedMoEPrepareAndFinalize | None:
         from .all2all_utils import maybe_make_prepare_finalize
 
-        return maybe_make_prepare_finalize(
-            moe_config or self.moe,
-            self.moe_quant_config,
-            routing_tables,
-            allow_new_interface=eep_stage,
-            use_monolithic=self.is_monolithic,
-            eep_stage=eep_stage,
+        pf = maybe_make_prepare_finalize(
+            self.moe, self.moe_quant_config, routing_tables
         )
-
-    def _make_eep_experts(
-        self,
-        source_experts: FusedMoEExpertsModular,
-        prepare_finalize: FusedMoEPrepareAndFinalizeModular,
-        moe_config: FusedMoEConfig,
-    ) -> FusedMoEExpertsModular:
-        experts_cls = source_experts.__class__
-        assert self.moe_quant_config is not None
-        experts_kwargs: dict[str, Any] = {
-            "moe_config": moe_config,
-            "quant_config": self.moe_quant_config,
-        }
-        if (
-            prepare_finalize.activation_format
-            == mk.FusedMoEActivationFormat.BatchedExperts
-        ):
-            max_num_tokens = prepare_finalize.max_num_tokens_per_rank()
-            assert max_num_tokens is not None
-            experts_kwargs.update(
-                max_num_tokens=max_num_tokens,
-                num_dispatchers=prepare_finalize.num_dispatchers(),
-            )
-
-        # Expert kernels with extra init params need explicit EEP support.
-        generic_arg_names = set(signature(mk.FusedMoEExperts.__init__).parameters)
-        ctor_arg_names = set(signature(experts_cls.__init__).parameters)
-        unsupported_args = ctor_arg_names - generic_arg_names
-        missing_args = set(experts_kwargs) - ctor_arg_names
-        if unsupported_args or missing_args:
-            raise NotImplementedError(
-                f"{experts_cls.__name__} experts do not support Elastic EP."
-            )
-
-        return experts_cls(**experts_kwargs)
-
-    def eep_make_staged_quant_method(
-        self,
-        _layer: torch.nn.Module,
-        moe: FusedMoEConfig,
-    ) -> "FusedMoEMethodBase | None":
-        if self.moe_kernel is None:
-            return None
-        if self.moe_kernel.is_monolithic:
-            raise NotImplementedError(
-                "Elastic EP full modular-kernel staging is not supported for "
-                "monolithic fused MoE kernels."
-            )
-        if self.moe_quant_config is None:
-            raise ValueError(
-                "Elastic EP full modular-kernel staging requires initialized "
-                "MoE quant config."
-            )
-
-        prepare_finalize = self._make_prepare_finalize(
-            routing_tables=None,
-            moe_config=moe,
-            eep_stage=True,
-        )
-        assert prepare_finalize is not None
-        assert isinstance(prepare_finalize, FusedMoEPrepareAndFinalizeModular)
-
-        source_experts = self.moe_kernel.fused_experts
-        assert isinstance(source_experts, FusedMoEExpertsModular)
-
-        experts = self._make_eep_experts(
-            source_experts,
-            prepare_finalize,
-            moe,
-        )
-
-        from .fused_moe_modular_method import FusedMoEModularMethod
-
-        if isinstance(self, FusedMoEModularMethod):
-            base_quant_method = self.old_quant_method
-        else:
-            base_quant_method = self
-        return FusedMoEModularMethod(
-            base_quant_method,
-            mk.FusedMoEKernel(
-                prepare_finalize,
-                experts,
-                inplace=self.moe_kernel.inplace,
-            ),
-        )
+        assert pf is None or isinstance(pf, FusedMoEPrepareAndFinalizeModular)
+        return pf
 
     def select_gemm_impl(
         self,
