+++ vllm/model_executor/layers/fused_moe/fused_moe_modular_method.py
@@ -6,7 +6,9 @@ import torch
 
 from vllm.logger import init_logger
 from vllm.model_executor.custom_op import CustomOp
-from vllm.model_executor.layers.fused_moe.config import FusedMoEQuantConfig
+from vllm.model_executor.layers.fused_moe.config import (
+    FusedMoEQuantConfig,
+)
 from vllm.model_executor.layers.fused_moe.fused_moe_method_base import (
     FusedMoEMethodBase,
 )
@@ -27,9 +29,7 @@ class FusedMoEModularMethod(FusedMoEMethodBase, CustomOp):
     # --8<-- [end:modular_fused_moe]
 
     def __init__(
-        self,
-        old_quant_method: FusedMoEMethodBase,
-        moe_kernel: FusedMoEKernel,
+        self, old_quant_method: FusedMoEMethodBase, moe_kernel: FusedMoEKernel
     ):
         super().__init__(moe_kernel.moe_config)
         self.moe_quant_config = old_quant_method.moe_quant_config
